fix(search-form): extract search actions component
