+++ src/components/loading-indicator/makeLoadable.tsx
@@ -6,10 +6,8 @@ export interface MakeLoadableProps {
     loadingIndicatorProps?: LoadingIndicatorProps;
 }
 
-const makeLoadable = <P extends MakeLoadableProps>(
-    BaseComponent: React.ComponentType<P>,
-): React.FC<P & MakeLoadableProps> => {
-    const LoadableComponent = ({ isLoading = false, loadingIndicatorProps = {}, ...rest }: MakeLoadableProps) =>
+const makeLoadable = <P extends object>(BaseComponent: React.ComponentType<P>): React.FC<P & MakeLoadableProps> => {
+    const LoadableComponent = ({ isLoading = false, loadingIndicatorProps = {}, ...rest }: P & MakeLoadableProps) =>
         isLoading ? <LoadingIndicator {...loadingIndicatorProps} /> : <BaseComponent {...(rest as P)} />;
 
     LoadableComponent.displayName = `Loadable${BaseComponent.displayName || BaseComponent.name || 'Component'}`;
