+++ src/components/search-form/messages.ts
@@ -1,0 +1,21 @@
+import { defineMessages } from 'react-intl';
+
+const messages = defineMessages({
+    clearButtonTitle: {
+        defaultMessage: 'Clear',
+        description: 'Title for a clear button',
+        id: 'boxui.searchForm.clearButtonTitle',
+    },
+    searchButtonTitle: {
+        defaultMessage: 'Search',
+        description: 'Title for a search button',
+        id: 'boxui.searchForm.searchButtonTitle',
+    },
+    searchLabel: {
+        defaultMessage: 'Search query',
+        description: 'Label for a search input',
+        id: 'boxui.searchForm.searchLabel',
+    },
+});
+
+export default messages;
